do not automatically attempt to create Google Sheets output
This commit cleans up a usability issue: that when you provide only your
github/gerrit credentials, this program fails because it tries to create google
sheets stuff (which is guaranteed to fail, since it hasn't been provided). It
fixes this by setting the default value of the sheets flag to empty, which
causes the program to exit before it tries to create sheets unless the user
specifically sets that flag.

Several examples on the README.md are therefore fixed with this.

It also fixes the problem a second time by setting tokenFile and credentialsFile
to default empty values, and adding a check that they actually got set by the
user before attempting to spin up a google sheet. In addition to fixing the
aforementioned problem (again), they are a little better user experience: users
won't get cryptic "file not found credentials.json" errors and have no idea what
is credentials.json.

Also cleans up:

- README.md 80 char limit
- README.md use multi-line code snippets for wide snippets
- Refactor google sheets ID parsing into its own function
- Use log.Print* everywhere rather than a mix of log.Print* and fmt.Print*